Everything goes down
This is the final and least polished version. I realize it'd be better to have the individual inputs instead of being concatenated values on one string variable, writings, as an array string[] writings, which then gets lumped together into a single string assigned to myWritings. This ultimately calls for a revised input mechanism that is worse and more vulnerable than the previous one it's replacing.